wujed_category_classifier/requirements.txt
fastapi
uvicorn
google-cloud-vision
openai
requests
pydantic
Pillow

